+++ CHANGELOG.md
@@ -7,6 +7,7 @@ V1.2
 * Improved removal stability: Attempting to remove more entries than available corrupted timesheet
 * Improved start command: Updating start-time of existing entry now accepts ID argument w/o "i=" prefix
 * Improved stop command: Updating end-time of existing entry now accepts ID argument w/o "i=" prefix
+* Improved task command: Added option to also append to entry comment while updating the task number 
 
 V1.1
 ----